requirements.txt
openai
# azure
azure-cognitiveservices-speech
praat-parselmouth
PyAudio
transformers
torch